pyproject.toml
[tool.poetry]
name = "bot-dev-container"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
ipykernel = "^6.17.0"
pymongo = "^4.3.2"
hyperliquid-python-sdk = "^0.9.0"
plotly = "^5.24.1"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
